+++ timezones/__init__.py
@@ -1,6 +1,6 @@
 """
-application init
+Application init
 """
 
-__version__ = "1.10.0"
+__version__ = "1.10.1"
 __title__ = "Time Zones"