+++ timezones/auth_hooks.py
@@ -1,5 +1,5 @@
 """
-hooks to AA
+Hooks to AA
 """
 
 # Django
@@ -15,10 +15,12 @@ from timezones.app_settings import allianceauth_discordbot_active
 
 
 class AaTimezonesMenuItem(MenuItemHook):  # pylint: disable=too-few-public-methods
-    """This class ensures only authorized users will see the menu entry"""
+    """
+    This class ensures only authorized users will see the menu entry
+    """
 
     def __init__(self):
-        # setup menu entry for sidebar
+        # Set up menu entry for sidebar
         MenuItemHook.__init__(
             self,
             _(__title__),
@@ -29,10 +31,11 @@ class AaTimezonesMenuItem(MenuItemHook):  # pylint: disable=too-few-public-metho
 
     def render(self, request):
         """
-        only if the user has access to this app
+        Only if the user has access to this app
         :param request:
         :return:
         """
+
         if request.user.has_perm("timezones.basic_access"):
             return MenuItemHook.render(self, request)
 
@@ -42,18 +45,20 @@ class AaTimezonesMenuItem(MenuItemHook):  # pylint: disable=too-few-public-metho
 @hooks.register("menu_item_hook")
 def register_menu():
     """
-    register our menu item
+    Register our menu item
     :return:
     """
+
     return AaTimezonesMenuItem()
 
 
 @hooks.register("url_hook")
 def register_urls():
     """
-    register our base url
+    Register our base url
     :return:
     """
+
     return UrlHook(urls, "timezones", r"^timezones/")